nxp-imx8: Adjust the unsupported list
Issue: LINUXEXEC-26017

GPU, VPU, NPU are placed in the unsupported list due to
licence restrictions, and comments are added, but this
is still ambiguous for non-developers. Although there
are comments, they are still incomprehensible. Someone
asked me whether we should tell customers which features
we support, based on PORTFOLIO-2485 or README.hardware.
If README.hardware is used, then these are placed in the
unsupported list, which is obviously contradictory, so
adjust these to the supported list and add comments.

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>